compile lib before calc stats
